% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkInputData}
\alias{checkInputData}
\title{Internal function: Validate input data for a spatpca object}
\usage{
checkInputData(x1, x2, Y1, Y2, M)
}
\arguments{
\item{x1}{Location matrix (\eqn{p \times d}) corresponding to Y1.}

\item{x2}{Location matrix (\eqn{q \times d}) corresponding to Y2.}

\item{Y1}{Data matrix (\eqn{n \times p}) of the first variable stores the values at \eqn{p} locations with sample size \eqn{n}.}

\item{Y2}{Data matrix (\eqn{n \times q}) of the second variable stores the values at \eqn{q} locations with sample size \eqn{n}.}

\item{M}{Number of folds for cross-validation}
}
\description{
Internal function: Validate input data for a spatpca object
}
\keyword{internal}
