% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plot_cv_field}
\alias{plot_cv_field}
\title{Internal function: Plot 2D fields for cross validation results}
\usage{
plot_cv_field(cv_data, variate)
}
\arguments{
\item{cv_data}{A dataframe contains columns \code{u}, \code{v}, and \code{cv}}

\item{variate}{A character represent the title}
}
\value{
A ggplot object
}
\description{
Internal function: Plot 2D fields for cross validation results
}
\keyword{internal}
