\name{SpatMCA-package}
\alias{SpatMCA-package}
\docType{package}
\title{Regularized Spatial Maximum Covariance Analysis}
\description{
A new regularization approach to estimate the leading coupled patterns via smoothness and sparseness penalties for spatial bivariate data that may be irregularly located in space.
}
\details{\tabular{ll}{
Package:\tab SpatMCA\cr
Type:\tab  Package\cr
Version:\tab   1.0.0.2\cr
Date:\tab   2017-03-17\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{ 
Wen-Ting Wang <egpivo@gmail.com> and Hsin-Cheng Huang <hchuang@stat.sinica.edu.tw>
}