% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatmcacvall_rcpp}
\alias{spatmcacvall_rcpp}
\title{Internal function: Overall M-fold Cross-validation of SpatMCA}
\usage{
spatmcacvall_rcpp(
  sxr,
  syr,
  Xr,
  Yr,
  M,
  K,
  tau1ur,
  tau2ur,
  tau1vr,
  tau2vr,
  nkr,
  maxit,
  tol,
  l2ur,
  l2vr
)
}
\arguments{
\item{sxr}{A location matrix for a variable Y}

\item{Yr}{A data matrix of X}

\item{M}{A number of folds}

\item{K}{The number of estimated eigenfunctions}

\item{tau1ur}{A range of tau1u}

\item{tau2ur}{A range of tau2u}

\item{tau1vr}{A range of tau1v}

\item{tau2vr}{A range of tau2v}

\item{nkr}{A vector of fold numbers}

\item{maxit}{A maximum number of iteration}

\item{tol}{A tolerance rate}

\item{l2ur}{A given tau2u}

\item{l2vr}{A given tau2v}
}
\value{
A list of selected parameters
}
\description{
Internal function: Overall M-fold Cross-validation of SpatMCA
}
\keyword{internal}
